Create new app arbovirus and moved virus model and tests from user app to arbovirus app
