+++ config/settings/base.py
@@ -85,6 +85,7 @@ THIRD_PARTY_APPS = [
 
 LOCAL_APPS = [
     "arboverse.users.apps.UsersConfig",
+    "arboverse.arbovirus.apps.ArbovirusConfig",
     # Your stuff: custom apps go here
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
